backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
PyMuPDF
python-docx
python-dotenv
